move initialization to lambda_handler due to before_first_request removal in flask 2.3.0
